smyrna: use a more appropriate type for 'ComboValuesCount'
Squashes two compiler warnings:

  gui/gui.c:94:96: warning: conversion to ‘size_t’ {aka ‘long unsigned
    int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
     94 |   attr[attrcount].ComboValuesCount - 1,
        |   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~
  gui/gui.c:95:78: warning: conversion to ‘size_t’ {aka ‘long unsigned
    int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
     95 |   attr[attrcount].ComboValuesCount,
        |   ~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~

+++ cmd/smyrna/smyrnadefs.h
@@ -25,6 +25,7 @@
 #include <glcomp/glcompset.h>
 #include "hier.h"
 #include <glcomp/glutils.h>
+#include <stddef.h>
 
 #ifdef	_WIN32			//this is needed on _WIN32 to get libglade see the callback
 #define _BB  __declspec(dllexport)
@@ -281,7 +282,7 @@ typedef struct
 	char *Default;
 	char Engine[GVK_FDP + 1];
 	char **ComboValues;
-	int ComboValuesCount;
+	size_t ComboValuesCount;
 	GtkWidget *attrWidget;
 
     } attribute;
